insert cpimage.py into the inheritance stack
