+++ py/legacypipe/bok.py
@@ -1,11 +1,11 @@
 import warnings
 
-from legacypipe.image import LegacySurveyImage
+from legacypipe.cpimage import CPImage
 
 '''
 Code specific to images from the 90prime camera on the Bok telescope.
 '''
-class BokImage(LegacySurveyImage):
+class BokImage(CPImage):
     '''
     Class for handling images from the 90prime camera processed by the
     NOAO Community Pipeline.
@@ -98,7 +98,7 @@ class BokImage(LegacySurveyImage):
         return self.remap_invvar_shotnoise(invvar, primhdr, img, dq)
 
     def remap_dq(self, dq, header):
-        from legacypipe.image import remap_dq_cp_codes
+        from legacypipe.cpimage import remap_dq_cp_codes
         # code 8: see https://github.com/legacysurvey/legacypipe/issues/645
         dq = remap_dq_cp_codes(dq, ignore_codes=[7, 8])
         return dq
