+++ py/legacypipe/mosaic.py
@@ -1,6 +1,6 @@
-from legacypipe.image import LegacySurveyImage
+from legacypipe.cpimage import CPImage
 
-class MosaicImage(LegacySurveyImage):
+class MosaicImage(CPImage):
     '''
     Class for handling images from the Mosaic3 camera processed by the
     NOAO Community Pipeline.
@@ -95,7 +95,7 @@ class MosaicImage(LegacySurveyImage):
         Called by get_tractor_image() to map the results from read_dq
         into a bitmask.
         '''
-        from legacypipe.image import remap_dq_cp_codes
+        from legacypipe.cpimage import remap_dq_cp_codes
         # code 8: https://github.com/legacysurvey/legacypipe/issues/644
         dq = remap_dq_cp_codes(dq, ignore_codes=[7, 8])
         return dq
